+++ packages/stg-evm-v2/devtools/config/mainnet/01/feelib-v1.usdt.config.ts
@@ -20,6 +20,7 @@ import {
     onMantle,
     onMetis,
     onOpt,
+    onPeaq,
     onPolygon,
     onRarible,
     onSei,
@@ -51,6 +52,7 @@ export default async (): Promise<OmniGraphHardhat<FeeLibV1NodeConfig, FeeLibV1Ed
     const mantleFeeLibV1 = onMantle(contract)
     const metisFeeLibV1 = onMetis(contract)
     const optFeeLibV1 = onOpt(contract)
+    const peaqFeeLibV1 = onPeaq(contract)
     const polygonFeeLibV1 = onPolygon(contract)
     const raribleFeeLibV1 = onRarible(contract)
     const seiFeeLibV1 = onSei(contract)
@@ -116,6 +118,10 @@ export default async (): Promise<OmniGraphHardhat<FeeLibV1NodeConfig, FeeLibV1Ed
             },
             {
                 contract: optFeeLibV1,
+                config: defaultNodeConfig,
+            },
+            {
+                contract: peaqFeeLibV1,
                 config: defaultNodeConfig,
             },
             {