+++ packages/stg-evm-v2/devtools/config/mainnet/01/usdc-admin.config.ts
@@ -1,13 +1,21 @@
+import assert from 'assert'
+
+import { TokenName } from '@stargatefinance/stg-definitions-v2'
 import { USDCNodeConfig } from '@stargatefinance/stg-devtools-v2'
 
 import { OmniGraphHardhat } from '@layerzerolabs/devtools-evm-hardhat'
 import { EndpointId } from '@layerzerolabs/lz-definitions'
 
 import { getUSDCProxyDeployName } from '../../../../ops/util'
+import { getAssetNetworkConfig } from '../../../../ts-src/utils/util'
 import { getSafeAddress } from '../../utils'
-import { onFlare, onGravity, onIota, onKlaytn, onLightlink, onRarible, onTaiko, onXchain } from '../utils'
+import { onFlare, onGravity, onIota, onKlaytn, onLightlink, onPeaq, onRarible, onTaiko, onXchain } from '../utils'
 
 const proxyContract = { contractName: getUSDCProxyDeployName() }
+
+// Except for Peaq where it's deployed externally
+const usdcPeaqAsset = getAssetNetworkConfig(EndpointId.PEAQ_V2_MAINNET, TokenName.USDC)
+assert(usdcPeaqAsset.address != null, `External USDC address not found for PEAQ`)
 
 export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
     // Get the corresponding underlying USDC contract
@@ -16,6 +24,7 @@ export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
     const iotaUSDC = onIota(proxyContract)
     const klaytnUSDC = onKlaytn(proxyContract)
     const lightlinkUSDC = onLightlink(proxyContract)
+    const peaqUSDC = onPeaq({ contractName: 'FiatTokenProxy', address: usdcPeaqAsset.address })
     const raribleUSDC = onRarible(proxyContract)
     const taikoUSDC = onTaiko(proxyContract)
     const xchainUSDC = onXchain(proxyContract)
@@ -25,6 +34,7 @@ export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
     const iotaStargateMultisig = getSafeAddress(EndpointId.IOTA_V2_MAINNET)
     const klaytnStargateMultisig = getSafeAddress(EndpointId.KLAYTN_V2_MAINNET)
     const lightlinkStargateMultisig = getSafeAddress(EndpointId.LIGHTLINK_V2_MAINNET)
+    const peaqStargateMultisig = getSafeAddress(EndpointId.PEAQ_V2_MAINNET)
     const raribleStargateMultisig = getSafeAddress(EndpointId.RARIBLE_V2_MAINNET)
     const taikoStargateMultisig = getSafeAddress(EndpointId.TAIKO_V2_MAINNET)
     const xchainStargateMultisig = getSafeAddress(EndpointId.XCHAIN_V2_MAINNET)
@@ -59,6 +69,12 @@ export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
                 contract: lightlinkUSDC,
                 config: {
                     admin: lightlinkStargateMultisig,
+                },
+            },
+            {
+                contract: peaqUSDC,
+                config: {
+                    admin: peaqStargateMultisig,
                 },
             },
             {