+++ packages/stg-definitions-v2/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stargatefinance/stg-definitions-v2
 
+## 1.3.0
+
+### Minor Changes
+
+- 676b2cd: Support for external USDT deployments
+
+## 1.2.5
+
+### Patch Changes
+
+- c8104b4: Peaq Deployments
+- 453dec7: peaq mainnet config
+
 ## 1.2.4
 
 ### Patch Changes